cleanup

+++ craft3data/src/com/hiveworkshop/wc3/gui/modeledit/BasicCoordinateSystem.java
@@ -22,22 +22,22 @@ public final class BasicCoordinateSystem implements CoordinateSystem {
 
 	@Override
 	public double convertX(final double x) {
-		return (x + cameraX) * zoom + width / 2;
+		return (x + cameraX) * zoom + (double) width / 2;
 	}
 
 	@Override
 	public double convertY(final double y) {
-		return ((-y + cameraY) * zoom) + height / 2;
+		return ((-y + cameraY) * zoom) + (double) height / 2;
 	}
 
 	@Override
 	public double geomX(final double x) {
-		return (x - width / 2) / zoom - cameraX;
+		return (x - (double) width / 2) / zoom - cameraX;
 	}
 
 	@Override
 	public double geomY(final double y) {
-		return -((y - height / 2) / zoom - cameraY);
+		return -((y - (double) height / 2) / zoom - cameraY);
 	}
 
 	@Override
